update smoothing tests
